working on collision
Added more to the scripts game and shape now shapes move down based on time.time if the controler dosn't move down than the shape does based off of time. Now im just trying make sure the system collides with the grid which its not doing yet but I just have to check all of the vars and make sure the values are set right. the gird is not based on the grid object and just the xy position.